ref(code split): separated client and constants functionality
