+++ llama-index-integrations/llms/llama-index-llms-deepinfra/llama_index/llms/deepinfra/constants.py
@@ -1,0 +1,10 @@
+"""DeepInfra API base URL."""
+API_BASE = "https://api.deepinfra.com/v1"
+"""DeepInfra Inference API endpoint."""
+INFERENCE_ENDPOINT = "inference"
+"""Chat API endpoint for DeepInfra."""
+CHAT_API_ENDPOINT = "openai/chat/completions"
+"""Environment variable name of DeepInfra API token."""
+ENV_VARIABLE = "DEEPINFRA_API_TOKEN"
+"""Default model name for DeepInfra embeddings."""
+DEFAULT_MODEL_NAME = "mistralai/Mixtral-8x22B-Instruct-v0.1"
